// src/generated/openapi/model/outputSchemaDTO.ts
/**
 * WTP API Starter
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OutputModelDTO } from './outputModelDTO';
import { OutputEnumDTO } from './outputEnumDTO';


export interface OutputSchemaDTO { 
    enums: Array<OutputEnumDTO>;
    models: Array<OutputModelDTO>;
}